backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
requests
email-validator
imap-tools
apscheduler
bcrypt
PyJWT
pymupdf
reportlab 
